scripts/requirements.txt
# Portfolio Optimization Scripts Dependencies
# Install with: pip install -r scripts/requirements.txt

# Image processing
Pillow>=10.0.0
